the start of changing hacks loading. i present you hash hacks loading. this is still crap but is has a shit ton more of potential then the current hacks and it can also be used to have a single hash/patch work on all system menu's. see hacks_hash.ini as to how the hacks.ini is made and how it works.
note that this code is still ugly as fuck and needs to be cleaned up.
i also want this to be the pramairy method of hacks loading since the plan is to improve it a shit ton more. sadly, some hashes are stuck to 1 system menu but even then you can just use 1 hash for 1 system menu version (no matter what the region is)

if you want to switch to the classic hacks loading, there is temp a setting to do that