Split up HBC & IOS Functions + Added OpenHBC support
Basically i got an email with the question for OpenHBC. i totally forgot
about that being a thing and just added support for it. for convience i
split up the IOS and HBC stuff so its easier to read, maintain and edit.
i should just make an array with all HBC titles and loop trough the
array but... meh xD

this is all untested (specially the stub part) but should work just fine
in theory.